Some changes

+++ app/Http/Controllers/ScoreController.php
@@ -65,6 +65,7 @@ class ScoreController extends Controller
      */
     public function edit(Score $score)
     {
+        $this->authorize('update', $score);
         $gimnastas = Gimnasta::all();
         $rounds = Round::all();
         $aparatos = Aparato::all();
@@ -77,6 +78,7 @@ class ScoreController extends Controller
      */
     public function update(Request $request, Score $score)
     {
+        $this->authorize('update', $score);
         $request->validate([
             'gimnastas_id'=>['required', 'exists:gimnastas,id'], 
             'events_id'=>['required', 'exists:events,id'],
@@ -99,6 +101,7 @@ class ScoreController extends Controller
      */
     public function destroy(Score $score)
     {
+        $this->authorize('delete', $score);
         $event=$score->events_id;
         $score->delete();
         return redirect()->route('event.show', $event);
